replace geoserver with SCIP API
