+++ docker/frontend.env
@@ -1,8 +1,7 @@
 REACT_APP_PCEX_API_URL=https://services.pacificclimate.org/dev/pcex/api
-REACT_APP_WATERSHED_TYPENAME=watersheds:watershed-degrees
+REACT_APP_SCIP_API_URL=https://services.pacificclimate.org/dev/scip/api/
 
 REACT_APP_REGIONS_SERVICE_URL=http://localhost/geoserver/watersheds/ows
 REACT_APP_BC_BASE_MAP_TILES_URL=https://services.pacificclimate.org/tiles/bc-albers-lite/{z}/{x}/{y}.png
 REACT_APP_BASE_MAP=BC
-REACT_APP_TEST_API_FILE=tasmax_mClimMean_BCCAQv2_PCIC12_historical-rcp85_rXi1p1_19610101-19901231_Canada
 PUBLIC_URL=http://localhost/frontend
