+++ src/components/AreaDisplay/AreaDisplay.js
@@ -1,24 +1,23 @@
-// AreaDisplay - interfaces with geoserver to get information about predefined areas
+// AreaDisplay - interfaces with the SCIP API to get information about predefined areas
 // (watersheds, conservation units, drainage basins, etc); lets the user select the
-// area they care about; displays categorical data about the selected area (accessed from
-//  geoserver).
+// area they care about; displays categorical data about the selected area 
 
-import {fetchWatersheds} from '../../data-services/regions.js';
+import {getWatersheds} from '../../data-services/scip-backend.js';
 import AreaSelector from '../AreaSelector/AreaSelector.js';
 import React, {useState, useEffect} from 'react';
 import {map, find} from 'lodash';
-import {parseRegion, filterRegions} from '../../helpers/RegionHelpers.js';
+import {parseRegions, filterRegions} from '../../helpers/GeographyHelpers.js';
 
 function AreaDisplay({onChangeRegion, region}) {
   const [regions, setRegions] = useState([]);
 
-  //fetch region list from geoserver.
+  //fetch region list from the API.
   // this only needs to be done once, when the component is loaded
   useEffect(() => {
     if(regions.length === 0) {
-        fetchWatersheds().then(
+        getWatersheds().then(
             data => {
-                setRegions(filterRegions(map( data.features, parseRegion)));
+                setRegions(filterRegions(parseRegions(data)));
             }
         );
     }
@@ -33,7 +32,6 @@ function AreaDisplay({onChangeRegion, region}) {
   return (
     <div className="AreaDisplay">
         <p>Currently Selected Region: {region ? region.name : "none"}</p>
-        {region ? <p>Drainage Area: {(region.area_meters / 1000000).toExponential(2)} km&sup2;</p> : ""}
         <AreaSelector
             regionNames={map(regions, 'name')}
             onChange={setRegion}
