+++ src/components/DailyDataDisplay/DailyDataDisplay.js
@@ -46,7 +46,7 @@ function DailyDataDisplay({region, rasterMetadata, model, emission}){
             });
         
         const api_calls = _.map(datafiles, datafile => {
-            return annualCycleDataRequest(region.geometry, datafile.file_id, 
+            return annualCycleDataRequest(region.boundary, datafile.file_id, 
                                    variable.representative.variable_id)
         });
         Promise.all(api_calls).then((api_responses)=> setDailyTimeSeries(api_responses));
