+++ src/components/MonthlyDataDisplay/MonthlyDataDisplay.js
@@ -34,7 +34,7 @@ function MonthlyDataDisplay({region, rasterMetadata, model, emission}){
         
         
         const api_calls = _.map(datafiles, datafile => {
-            return annualCycleDataRequest(region.geometry, datafile.file_id, 
+            return annualCycleDataRequest(region.boundary, datafile.file_id, 
                                    variable.representative.variable_id)
         });        
         Promise.all(api_calls).then((api_responses)=> setAnnualCycleTimeSeries(api_responses));
