+++ src/components/YearlyDataDisplay/YearlyDataDisplay.js
@@ -24,7 +24,7 @@ function YearlyDataDisplay({region, rasterMetadata, model, emission}){
   
     useEffect(() => {
       if(region && variable) {
-        longTermAverageDataRequest(region.geometry, 
+        longTermAverageDataRequest(region.boundary, 
                                    variable.representative.variable_id,
                                    model,
                                    emission,
